+++ src/modules/burpread_mod.f90
@@ -5032,11 +5032,19 @@ CONTAINS
 
         call burp_init_report_write(inputFile, copyReport, iostat=error)
 
+        call burp_get_property(inputReport, stnid = station_id)
+
         ! loop on blocks
         ! --------------------
         ref_blk = 0
       
         blocks: do
+          ! We skip the writing of the block in the resume record
+          ! because writing it may lead to 'floating point exception'
+          ! with 'mrbcvt' as called inside 'burp_write_block'.
+          if (station_id(1:2) == ">>") then
+            exit blocks
+          end if
 
           ref_blk = burp_find_block(inputReport, &
                block       = inputBlock,         &