+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+ * Avoid writing any block from the resume record in the routine `brpr_addElementsToBurp` (#868 and !770)
  * Fix some potential out of bounds array access in midas-obsSelection and midas-genCoeff programs (#701 and !647)
 
 ### Removed